renamed name_validation to domname
