+++ components/script/dom/bindings/mod.rs
@@ -139,13 +139,13 @@ pub(crate) mod buffer_source;
 pub(crate) mod cell;
 pub(crate) mod constructor;
 pub(crate) mod conversions;
+pub(crate) mod domname;
 pub(crate) mod error;
 pub(crate) mod frozenarray;
 pub(crate) mod function;
 pub(crate) mod import;
 pub(crate) mod inheritance;
 pub(crate) mod like;
-pub(crate) mod name_validation;
 pub(crate) mod principals;
 pub(crate) mod proxyhandler;
 pub(crate) mod refcounted;
